refactor(direct-control): extract runtime inspection constants
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only the runtime inspection default App UI/Tuner API root catalogs and bounded root maxKeys/maxMethods defaults into src/runtime/inspection-constants.ts while keeping public facade re-exports and injected runtime dependencies in index.ts.

Preserves runtime API default-root selection, bounded root cap defaults, public constant values, and debug/internal inspection classification. Broader public constants/types, procedure schemas, operation/proof telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- public-api.test.ts runtime-and-catalog.test.ts
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- git diff --check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- ownership/no-cycle scans for runtime inspection constants and executable facade imports

Local package/constants relocation proof only; no runtime, live-game, AI/hotseat product-path, semantic CLI, telemetry, runtime reflection authority, or procedure-core proof claimed.
