+++ packages/civ7-direct-control/src/runtime/inspection-constants.ts
@@ -1,0 +1,25 @@
+export const DEFAULT_CIV7_APP_UI_API_ROOTS = [
+  "Network",
+  "Configuration",
+  "GameSetup",
+  "Autoplay",
+  "Game",
+  "UI",
+  "GameContext",
+  "PlayerIds",
+  "Players",
+  "GameplayMap",
+] as const;
+
+export const DEFAULT_CIV7_TUNER_API_ROOTS = [
+  "Game",
+  "Autoplay",
+  "Players",
+  "GameplayMap",
+  "ResourceBuilder",
+  "GameInfo",
+  "PlayerIds",
+] as const;
+
+export const DEFAULT_CIV7_ROOT_MAX_KEYS = 100;
+export const DEFAULT_CIV7_ROOT_MAX_METHODS = 100;
